Increase TWAI rx/tx queue lengths
If the OpenCAN tx scheduler sends more than like 8 messages at once,
they're likely to get dropped, because it's queueing them up much
faster than they're being sent. This behavior was observed on
the motorsports AMS ECU.

Increase the queue lengths to 64. In the future, the TX scheduler should
be made more intelligent so that it staggers the transmission of cyclics.

64 is an arbitrary upper limit increase; it's not enough to fully solve
the problem.